backend/flask_auth/requirements.txt
Flask==2.3.4
flask-cors==3.0.10
bcrypt==4.0.1
PyJWT==2.8.0
